+++ tomcat/src/main/java/nextstep/jwp/handler/LoginHandler.java
@@ -3,13 +3,13 @@ package nextstep.jwp.handler;
 import java.util.Optional;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.Cookies;
-import org.apache.coyote.http11.FileHandler;
-import org.apache.coyote.http11.FormData;
+import org.apache.coyote.http11.header.Cookies;
+import org.apache.coyote.http11.handler.FileHandler;
+import org.apache.coyote.http11.body.FormData;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.Session;
-import org.apache.coyote.http11.SessionManager;
+import org.apache.coyote.http11.session.Session;
+import org.apache.coyote.http11.session.SessionManager;
 
 public class LoginHandler extends FileHandler {
