+++ tomcat/src/main/java/nextstep/jwp/handler/RegisterHandler.java
@@ -2,8 +2,8 @@ package nextstep.jwp.handler;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.FileHandler;
-import org.apache.coyote.http11.FormData;
+import org.apache.coyote.http11.handler.FileHandler;
+import org.apache.coyote.http11.body.FormData;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
