--- tomcat/src/main/java/org/apache/coyote/http11/Handler.java
@@ -1,6 +1,0 @@
-package org.apache.coyote.http11;
-
-public interface Handler {
-
-    HttpResponse handle(HttpRequest httpRequest);
-}
