+++ tomcat/src/main/java/org/apache/coyote/http11/handler/FileHandler.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler;
 
 import java.io.File;
 import java.io.IOException;
@@ -6,6 +6,8 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.util.Map;
 import java.util.Objects;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
 
 public class FileHandler implements Handler {
