+++ tomcat/src/main/java/org/apache/coyote/http11/handler/Handler.java
@@ -1,0 +1,9 @@
+package org.apache.coyote.http11.handler;
+
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+
+public interface Handler {
+
+    HttpResponse handle(HttpRequest httpRequest);
+}
