+++ tomcat/src/main/java/org/apache/coyote/http11/header/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.header;
 
 public class ContentType extends HttpHeader {
